docs: add response entities and annotations to avoid recursion

+++ pizzeria_webapi/src/main/java/org/lessons/pizzeria/webapi/pizzeria_webapi/controller/PizzaRestController.java
@@ -1,10 +1,13 @@
 package org.lessons.pizzeria.webapi.pizzeria_webapi.controller;
 
 import java.util.List;
+import java.util.Optional;
 
 import org.lessons.pizzeria.webapi.pizzeria_webapi.model.Pizza;
 import org.lessons.pizzeria.webapi.pizzeria_webapi.service.PizzaService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatusCode;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -30,27 +33,45 @@ public class PizzaRestController {
 
     // show
     @GetMapping("/{id}")
-    public Pizza show(@PathVariable int id) {
-        Pizza pizza = pizzaService.getById(id);
-        return pizza;
+    public ResponseEntity<Pizza> show(@PathVariable int id) {
+        Optional<Pizza> pizzaAttempt = pizzaService.findById(id);
+
+        if (pizzaAttempt.isEmpty()) {
+            return new ResponseEntity<Pizza>(HttpStatusCode.valueOf(404));
+        }
+
+        return new ResponseEntity<Pizza>(pizzaAttempt.get(), HttpStatusCode.valueOf(200));
     }
 
     // create
     @PostMapping
-    public Pizza store(@RequestBody Pizza pizza) {
-        return pizzaService.create(pizza);
+    public ResponseEntity<Pizza> store(@RequestBody Pizza pizza) {
+        return new ResponseEntity<Pizza>(pizzaService.create(pizza), HttpStatusCode.valueOf(201));
     }
 
     // update
     @PutMapping("/{id}")
-    public Pizza update(@PathVariable int id, @RequestBody Pizza pizza) {
+    public ResponseEntity<Pizza> update(@PathVariable int id, @RequestBody Pizza pizza) {
+        Optional<Pizza> pizzaAttempt = pizzaService.findById(id);
+
+        if (pizzaAttempt.isEmpty()) {
+            return new ResponseEntity<Pizza>(HttpStatusCode.valueOf(404));
+        }
+
         pizza.setId(id);
-        return pizzaService.update(pizza);
+        return new ResponseEntity<Pizza>(pizzaService.update(pizza), HttpStatusCode.valueOf(201));
     }
 
     // delete
     @DeleteMapping("/{id}")
-    public void delete(@PathVariable int id) {
+    public ResponseEntity<Pizza> delete(@PathVariable int id) {
+        Optional<Pizza> pizzaAttempt = pizzaService.findById(id);
+
+        if (pizzaAttempt.isEmpty()) {
+            return new ResponseEntity<Pizza>(HttpStatusCode.valueOf(404));
+        }
+
         pizzaService.deleteById(id);
+        return new ResponseEntity<Pizza>(HttpStatusCode.valueOf(204));
     }
 }
